Moved GenerateHash to its own file
