+++ pkg/controller/util/dispatcher.go
@@ -1,7 +1,6 @@
 package util
 
 import (
-	"crypto/md5"
 	"fmt"
 	kafkav1beta1 "knative.dev/eventing-contrib/kafka/channel/pkg/apis/messaging/v1beta1"
 )
@@ -17,10 +16,4 @@ func DispatcherDnsSafeName(channel *kafkav1beta1.KafkaChannel) string {
 	safeChannelNamespace := GenerateValidDnsName(channel.Namespace, 16, false, false)
 	hash := GenerateHash(channel.Name + channel.Namespace, 8)
 	return fmt.Sprintf("%s-%s-%s-dispatcher", safeChannelName, safeChannelNamespace, hash)
-}
-
-// Generate An MD5 Hash Of A String And Return Desired Number Of Characters
-func GenerateHash(stringToHash string, length int) string {
-	// Create an MD5 hash and return however many characters the caller wants (note that the max is 32 for MD5)
-	return fmt.Sprintf(fmt.Sprintf("%%.%ds", length), fmt.Sprintf("%x", md5.Sum([]byte(stringToHash))))
 }
