+++ pkg/controller/util/dispatcher_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"testing"
 
-	"crypto/md5"
 	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	kafkav1beta1 "knative.dev/eventing-contrib/kafka/channel/pkg/apis/messaging/v1beta1"
@@ -140,30 +139,4 @@ func TestDispatcherDnsSafeName_LongNamesDifferent(t *testing.T) {
 		assert.Equal(t, expectedResult2, actualResult2)
 		assert.NotEqual(t, actualResult1, actualResult2)
 	}
-}
-
-// Test the GenerateHash Functionality
-func TestGenerateHash(t *testing.T) {
-	// Define The TestCase Struct
-	type TestCase struct {
-		Name   string
-		Length int
-	}
-
-	// Create The TestCases
-	testCases := []TestCase{
-		{Name: "", Length: 32},
-		{Name: "short string", Length: 4},
-		{Name: "long string, 8-character hash", Length: 8},
-		{Name: "odd hash length, 13-characters", Length: 13},
-		{Name: "very long string with 16-character hash and more than 64 characters total", Length: 16},
-	}
-
-	// Run The TestCases
-	for _, testCase := range testCases {
-		hash := GenerateHash(testCase.Name, testCase.Length)
-		expected := fmt.Sprintf("%x", md5.Sum([]byte(testCase.Name)))[0:testCase.Length]
-		assert.Equal(t, expected, hash)
-	}
-
 }
